requirements.txt
mysql-connector-python==9.1.0
Pillow==11.0.0
tkcalendar==1.6.1
python-dotenv==1.0.1
pytest==8.3.4
